+++ Tracker/Helpers/UserDefaultsBacked.swift
@@ -11,12 +11,8 @@ struct UserDefaultsBacked<Value> {
     var storage: UserDefaults = .standard
 
     var wrappedValue: Value {
-        get {
-            return storage.object(forKey: key) as? Value ?? defaultValue
-        }
-        set {
-            storage.set(newValue, forKey: key)
-        }
+        get { storage.object(forKey: key) as? Value ?? defaultValue }
+        set { storage.set(newValue, forKey: key) }
     }
 }
