+++ Tracker/Screens/OnBoarding/Controllers/OnboardingViewController.swift
@@ -10,9 +10,9 @@ final class OnboardingViewController: UIPageViewController, UIPageViewController
 
     private let firstPageOnboardingView = OnboardingView()
     private let secondPageOnboardingView = OnboardingView()
-    var pages: [UIViewController] = []
+    private var pages: [UIViewController] = []
 
-    lazy var pageControl: UIPageControl = {
+    private lazy var pageControl: UIPageControl = {
         let pageControl = UIPageControl()
         pageControl.numberOfPages = pages.count
         pageControl.currentPage = 0
